curve: Consolidate the vector code
A massive explosion of files is super annoying, though I still wish I
had a `pub(crate)` equivalent.  And no, `internal/` doesn't cut it.
