+++ curve/edwards_vector_amd64.s
