+++ curve/scalar_mul.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2016-2019 Isis Agora Lovecruft, Henry de Valence. All rights reserved.
-// Copyright (c) 2020-2021 Oasis Labs Inc.  All rights reserved.
+// Copyright (c) 2021 Oasis Labs Inc.  All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are
@@ -28,55 +28,60 @@
 // NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-// +build !amd64 purego forcenoasm force32bit
-
 package curve
 
 import "github.com/oasisprotocol/curve25519-voi/curve/scalar"
 
-// EdwardsBasepointTable defines a precomputed table of multiples of a
-// basepoint, for accelerating fixed-based scalar multiplication.
-type EdwardsBasepointTable struct {
-	inner *edwardsBasepointTableGeneric
-}
-
-func (tbl *EdwardsBasepointTable) mul(out *EdwardsPoint, scalar *scalar.Scalar) *EdwardsPoint {
-	return tbl.inner.Mul(out, scalar)
-}
-
-// Basepoint returns the basepoint of the table.
-func (tbl *EdwardsBasepointTable) Basepoint() *EdwardsPoint {
-	return tbl.inner.Basepoint()
-}
-
-// NewEdwardsBasepointTable creates a table of precomputed multiples of
-// `basepoint`.
-func NewEdwardsBasepointTable(basepoint *EdwardsPoint) *EdwardsBasepointTable {
-	return &EdwardsBasepointTable{
-		inner: newEdwardsBasepointTableGeneric(basepoint),
+func edwardsMul(out, point *EdwardsPoint, scalar *scalar.Scalar) *EdwardsPoint {
+	switch supportsVectorizedEdwards {
+	case true:
+		return edwardsMulVector(out, point, scalar)
+	default:
+		return edwardsMulGeneric(out, point, scalar)
 	}
 }
 
-func edwardsMul(out, point *EdwardsPoint, scalar *scalar.Scalar) *EdwardsPoint {
-	return edwardsMulGeneric(out, point, scalar)
-}
-
 func edwardsDoubleScalarMulBasepointVartime(out *EdwardsPoint, a *scalar.Scalar, A *EdwardsPoint, b *scalar.Scalar) *EdwardsPoint {
-	return edwardsDoubleScalarMulBasepointVartimeGeneric(out, a, A, b)
+	switch supportsVectorizedEdwards {
+	case true:
+		return edwardsDoubleScalarMulBasepointVartimeVector(out, a, A, b)
+	default:
+		return edwardsDoubleScalarMulBasepointVartimeGeneric(out, a, A, b)
+	}
 }
 
 func edwardsMulAbglsvPorninVartime(out *EdwardsPoint, a *scalar.Scalar, A *EdwardsPoint, b *scalar.Scalar, C *EdwardsPoint) *EdwardsPoint {
-	return edwardsMulAbglsvPorninVartimeGeneric(out, a, A, b, C)
+	switch supportsVectorizedEdwards {
+	case true:
+		return edwardsMulAbglsvPorninVartimeVector(out, a, A, b, C)
+	default:
+		return edwardsMulAbglsvPorninVartimeGeneric(out, a, A, b, C)
+	}
 }
 
 func edwardsMultiscalarMulStraus(out *EdwardsPoint, scalars []*scalar.Scalar, points []*EdwardsPoint) *EdwardsPoint {
-	return edwardsMultiscalarMulStrausGeneric(out, scalars, points)
+	switch supportsVectorizedEdwards {
+	case true:
+		return edwardsMultiscalarMulStrausVector(out, scalars, points)
+	default:
+		return edwardsMultiscalarMulStrausGeneric(out, scalars, points)
+	}
 }
 
 func edwardsMultiscalarMulStrausVartime(out *EdwardsPoint, scalars []*scalar.Scalar, points []*EdwardsPoint) *EdwardsPoint {
-	return edwardsMultiscalarMulStrausVartimeGeneric(out, scalars, points)
+	switch supportsVectorizedEdwards {
+	case true:
+		return edwardsMultiscalarMulStrausVartimeVector(out, scalars, points)
+	default:
+		return edwardsMultiscalarMulStrausVartimeGeneric(out, scalars, points)
+	}
 }
 
 func edwardsMultiscalarMulPippengerVartime(out *EdwardsPoint, scalars []*scalar.Scalar, points []*EdwardsPoint) *EdwardsPoint {
-	return edwardsMultiscalarMulPippengerVartimeGeneric(out, scalars, points)
+	switch supportsVectorizedEdwards {
+	case true:
+		return edwardsMultiscalarMulPippengerVartimeVector(out, scalars, points)
+	default:
+		return edwardsMultiscalarMulPippengerVartimeGeneric(out, scalars, points)
+	}
 }
