+++ curve/scalar_mul_variable_base.go
@@ -75,3 +75,33 @@ func edwardsMulGeneric(out, point *EdwardsPoint, scalar *scalar.Scalar) *Edwards
 	}
 	return out.setCompleted(&tmp1)
 }
+
+func edwardsMulVector(out, point *EdwardsPoint, scalar *scalar.Scalar) *EdwardsPoint {
+	// Construct a lookup table of [P,2P,3P,4P,5P,6P,7P,8P]
+	lookupTable := newCachedPointLookupTable(point)
+
+	// Setting s = scalar, compute
+	//
+	//    s = s_0 + s_1*16^1 + ... + s_63*16^63,
+	//
+	// with `-8 ≤ s_i < 8` for `0 ≤ i < 63` and `-8 ≤ s_63 ≤ 8`.
+	scalarDigits := scalar.ToRadix16()
+	// Compute s*P as
+	//
+	//    s*P = P*(s_0 +   s_1*16^1 +   s_2*16^2 + ... +   s_63*16^63)
+	//    s*P =  P*s_0 + P*s_1*16^1 + P*s_2*16^2 + ... + P*s_63*16^63
+	//    s*P = P*s_0 + 16*(P*s_1 + 16*(P*s_2 + 16*( ... + P*s_63)...))
+	//
+	// We sum right-to-left.
+	var (
+		q   extendedPoint
+		tmp cachedPoint
+	)
+	q.Identity()
+	for i := 63; i >= 0; i-- {
+		q.MulByPow2(&q, 4)
+		tmp = lookupTable.Lookup(scalarDigits[i])
+		q.AddExtendedCached(&q, &tmp)
+	}
+	return out.setExtended(&q)
+}
