+++ curve/window.go
@@ -125,6 +125,45 @@ func newAffineNielsPointLookupTable(ep *EdwardsPoint) affineNielsPointLookupTabl
 	return affineNielsPointLookupTable(points)
 }
 
+type cachedPointLookupTable [8]cachedPoint
+
+func (tbl *cachedPointLookupTable) Lookup(x int8) cachedPoint {
+	// Compute xabs = |x|
+	xmask := x >> 7
+	xabs := uint8((x + xmask) ^ xmask)
+
+	// Set t = 0 * P = identity
+	var t cachedPoint
+	lookupCached(tbl, &t, xabs)
+	// Now t == |x| * P.
+
+	negMask := int(byte(xmask & 1))
+	t.ConditionalNegate(negMask)
+	// Now t == x * P.
+
+	return t
+}
+
+func newCachedPointLookupTable(ep *EdwardsPoint) cachedPointLookupTable {
+	var (
+		epExtended extendedPoint
+		epCached   cachedPoint
+	)
+
+	epCached.SetExtended(epExtended.SetEdwards(ep))
+
+	points := [8]cachedPoint{
+		epCached, epCached, epCached, epCached,
+		epCached, epCached, epCached, epCached,
+	}
+	for i := 0; i < 7; i++ {
+		var tmp extendedPoint
+		points[i+1].SetExtended(tmp.AddExtendedCached(&epExtended, &points[i]))
+	}
+
+	return cachedPointLookupTable(points)
+}
+
 // Holds odd multiples 1A, 3A, ..., 15A of a point A.
 type projectiveNielsPointNafLookupTable [8]projectiveNielsPoint
 
@@ -152,6 +191,36 @@ func newProjectiveNielsPointNafLookupTable(ep *EdwardsPoint) projectiveNielsPoin
 	}
 
 	return projectiveNielsPointNafLookupTable(Ai)
+}
+
+// Holds odd multiples 1A, 3A, ..., 15A of a point A.
+type cachedPointNafLookupTable [8]cachedPoint
+
+func (tbl *cachedPointNafLookupTable) Lookup(x uint8) *cachedPoint {
+	return &tbl[x/2]
+}
+
+func newCachedPointNafLookupTable(ep *EdwardsPoint) cachedPointNafLookupTable {
+	var (
+		epExtended extendedPoint
+		epCached   cachedPoint
+		A2         extendedPoint
+	)
+
+	epCached.SetExtended(epExtended.SetEdwards(ep))
+
+	Ai := [8]cachedPoint{
+		epCached, epCached, epCached, epCached,
+		epCached, epCached, epCached, epCached,
+	}
+
+	A2.Double(&epExtended)
+	for i := 0; i < 7; i++ {
+		var tmp extendedPoint
+		Ai[i+1].SetExtended(tmp.AddExtendedCached(&A2, &Ai[i]))
+	}
+
+	return cachedPointNafLookupTable(Ai)
 }
 
 // Holds stuff up to 8.
@@ -191,6 +260,48 @@ func newAffineNielsPointShl128NafLookupTable(ep *EdwardsPoint) affineNielsPointN
 		tmp.setAffineNiels(&ap)
 		tmp.mulByPow2(&tmp, 128)
 		table[i].SetEdwards(&tmp)
+	}
+
+	return table
+}
+
+// Holds stuff up to 8.
+type cachedPointNafLookupTable8 [64]cachedPoint
+
+func (tbl *cachedPointNafLookupTable8) Lookup(x uint8) *cachedPoint {
+	return &tbl[x/2]
+}
+
+func newCachedPointNafLookupTable8(ep *EdwardsPoint) cachedPointNafLookupTable8 { //nolint:unused,deadcode
+	var (
+		epExtended extendedPoint
+		epCached   cachedPoint
+		A2         extendedPoint
+	)
+
+	epCached.SetExtended(epExtended.SetEdwards(ep))
+
+	var Ai [64]cachedPoint
+	for i := range Ai {
+		Ai[i] = epCached
+	}
+
+	A2.Double(&epExtended)
+	for i := 0; i < 63; i++ {
+		var tmp extendedPoint
+		Ai[i+1].SetExtended(tmp.AddExtendedCached(&A2, &Ai[i]))
+	}
+
+	return cachedPointNafLookupTable8(Ai)
+}
+
+func newCachedPointShl128NafLookupTable8(ep *EdwardsPoint) cachedPointNafLookupTable8 { //nolint:unused,deadcode
+	table := newCachedPointNafLookupTable8(ep)
+	for i, cp := range table {
+		var tmp extendedPoint
+		tmp.AddExtendedCached(tmp.Identity(), &cp)
+		tmp.MulByPow2(&tmp, 128)
+		table[i].SetExtended(&tmp)
 	}
 
 	return table
